requirements.txt
playwright>=1.47.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
extruct>=0.17.0
tldextract>=5.1.0
w3lib>=2.2.0
click>=8.1.0
python-docx>=1.1.0
